raspberrypi-compagnion/requirements.txt
certifi==2020.12.5
chardet==4.0.0
gevent==21.1.2
greenlet==1.1.0
idna==2.10
numpy==1.19.5
opencv-python==4.5.2.52
pigpio==1.78
pyserial==3.5
python-engineio==3.13.2
python-engineio-client==1.1
python-socketio==4.6.1
python-socketio-client==1.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
websocket-client==1.0.0
WiiProxy==2.0
zope.event==4.5.0
zope.interface==5.4.0
